codeforces/cpp/1900-1999/G/1915G.cp

+++ codeforces/cpp/1900-1999/G/1915G.cpp
@@ -11,20 +11,20 @@ constexpr ll inf = 1e18;
 
 auto solve()
 {
-    int n, m;
+    ll n, m;
     cin >> n >> m;
-    int u, v, w;
+    ll u, v, w;
     vector<vector<pair<int, int>>> adj(n + 1);
     vector<ll> slow(n + 1);
 
-    for (int i = 0; i < m; ++i)
+    for (ll i = 0; i < m; ++i)
     {
         cin >> u >> v >> w;
         adj[u].emplace_back(v, w);
         adj[v].emplace_back(u, w);
     }
 
-    for (int i = 1; i <= n; ++i) cin >> slow[i];
+    for (ll i = 1; i <= n; ++i) cin >> slow[i];
     priority_queue<tuple<ll, ll, ll>> q;
     q.emplace(0LL, 1LL, slow[1]);
